+++ product/login/src/main/java/io/codekvast/login/heroku/HerokuSsoController.java
@@ -23,6 +23,7 @@ package io.codekvast.login.heroku;
 
 import io.codekvast.common.security.SecurityService;
 import io.codekvast.login.bootstrap.CodekvastLoginSettings;
+import io.codekvast.login.metrics.MetricsService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
@@ -51,6 +52,7 @@ public class HerokuSsoController {
 
     private final CodekvastLoginSettings settings;
     private final SecurityService securityService;
+    private final MetricsService metricsService;
 
     @ExceptionHandler
     public ResponseEntity<String> onAuthenticationException(AuthenticationException e) {
@@ -74,6 +76,7 @@ public class HerokuSsoController {
         response.setStatus(HttpStatus.TEMPORARY_REDIRECT.value());
         response
             .setHeader(HttpHeaders.LOCATION, String.format("%s/dashboard/heroku/sso/%s/%s", settings.getDashboardBaseUrl(), code, navData));
+        metricsService.countLogin("heroku");
     }
 
 }